Add files via upload
Created a simple website where payment gateway is integrated. There will be a simple donate button on homepage. On clicking the donate button, the user will land on the payment page where user can select the amount to be paid and the payment type, e.g. credit card, Paypal, etc. Once the payment is done and invoice will be generated and email will be sent to the user for the payment received. The invoice will contain the amount. My website name is Sahayak Foundation and it's for the donation for the help of needy people, trees, against pollution, etc.